+++ api-client-java/src/main/java/io/daytona/api/client/model/PaginatedAuditLogs.java
@@ -21,7 +21,6 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import io.daytona.api.client.model.AuditLog;
 import java.io.IOException;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -62,17 +61,17 @@ public class PaginatedAuditLogs {
   public static final String SERIALIZED_NAME_TOTAL = "total";
   @SerializedName(SERIALIZED_NAME_TOTAL)
   @javax.annotation.Nonnull
-  private BigDecimal total;
+  private Integer total;
 
   public static final String SERIALIZED_NAME_PAGE = "page";
   @SerializedName(SERIALIZED_NAME_PAGE)
   @javax.annotation.Nonnull
-  private BigDecimal page;
+  private Integer page;
 
   public static final String SERIALIZED_NAME_TOTAL_PAGES = "totalPages";
   @SerializedName(SERIALIZED_NAME_TOTAL_PAGES)
   @javax.annotation.Nonnull
-  private BigDecimal totalPages;
+  private Integer totalPages;
 
   public static final String SERIALIZED_NAME_NEXT_TOKEN = "nextToken";
   @SerializedName(SERIALIZED_NAME_NEXT_TOKEN)
@@ -109,7 +108,7 @@ public class PaginatedAuditLogs {
   }
 
 
-  public PaginatedAuditLogs total(@javax.annotation.Nonnull BigDecimal total) {
+  public PaginatedAuditLogs total(@javax.annotation.Nonnull Integer total) {
     this.total = total;
     return this;
   }
@@ -119,16 +118,16 @@ public class PaginatedAuditLogs {
    * @return total
    */
   @javax.annotation.Nonnull
-  public BigDecimal getTotal() {
+  public Integer getTotal() {
     return total;
   }
 
-  public void setTotal(@javax.annotation.Nonnull BigDecimal total) {
+  public void setTotal(@javax.annotation.Nonnull Integer total) {
     this.total = total;
   }
 
 
-  public PaginatedAuditLogs page(@javax.annotation.Nonnull BigDecimal page) {
+  public PaginatedAuditLogs page(@javax.annotation.Nonnull Integer page) {
     this.page = page;
     return this;
   }
@@ -138,16 +137,16 @@ public class PaginatedAuditLogs {
    * @return page
    */
   @javax.annotation.Nonnull
-  public BigDecimal getPage() {
+  public Integer getPage() {
     return page;
   }
 
-  public void setPage(@javax.annotation.Nonnull BigDecimal page) {
+  public void setPage(@javax.annotation.Nonnull Integer page) {
     this.page = page;
   }
 
 
-  public PaginatedAuditLogs totalPages(@javax.annotation.Nonnull BigDecimal totalPages) {
+  public PaginatedAuditLogs totalPages(@javax.annotation.Nonnull Integer totalPages) {
     this.totalPages = totalPages;
     return this;
   }
@@ -157,11 +156,11 @@ public class PaginatedAuditLogs {
    * @return totalPages
    */
   @javax.annotation.Nonnull
-  public BigDecimal getTotalPages() {
+  public Integer getTotalPages() {
     return totalPages;
   }
 
-  public void setTotalPages(@javax.annotation.Nonnull BigDecimal totalPages) {
+  public void setTotalPages(@javax.annotation.Nonnull Integer totalPages) {
     this.totalPages = totalPages;
   }
