+++ api-client-ruby/lib/daytona_api_client/models/paginated_traces.rb
@@ -51,9 +51,9 @@ module DaytonaApiClient
     def self.openapi_types
       {
         :'items' => :'Array<TraceSummary>',
-        :'total' => :'Float',
-        :'page' => :'Float',
-        :'total_pages' => :'Float'
+        :'total' => :'Integer',
+        :'page' => :'Integer',
+        :'total_pages' => :'Integer'
       }
     end
